+++ sample/src/main/java/io/noties/markwon/sample/MainActivity.java
@@ -30,8 +30,10 @@ import io.noties.markwon.sample.inlineparser.InlineParserActivity;
 import io.noties.markwon.sample.latex.LatexActivity;
 import io.noties.markwon.sample.notification.NotificationActivity;
 import io.noties.markwon.sample.precomputed.PrecomputedActivity;
+import io.noties.markwon.sample.precomputed.PrecomputedFutureActivity;
 import io.noties.markwon.sample.recycler.RecyclerActivity;
 import io.noties.markwon.sample.simpleext.SimpleExtActivity;
+import io.noties.markwon.sample.table.TableActivity;
 import io.noties.markwon.sample.tasklist.TaskListActivity;
 
 public class MainActivity extends Activity {
@@ -123,6 +125,10 @@ public class MainActivity extends Activity {
                 activity = PrecomputedActivity.class;
                 break;
 
+            case PRECOMPUTED_FUTURE_TEXT:
+                activity = PrecomputedFutureActivity.class;
+                break;
+
             case EDITOR:
                 activity = EditorActivity.class;
                 break;
@@ -145,6 +151,10 @@ public class MainActivity extends Activity {
 
             case REMOTE_VIEWS:
                 activity = NotificationActivity.class;
+                break;
+
+            case TABLE:
+                activity = TableActivity.class;
                 break;
 
             default: