+++ sample/src/main/java/io/noties/markwon/sample/basicplugins/BasicPluginsActivity.java
@@ -3,11 +3,8 @@ package io.noties.markwon.sample.basicplugins;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
-import android.text.Spannable;
-import android.text.Spanned;
 import android.text.TextUtils;
 import android.text.style.ForegroundColorSpan;
-import android.view.View;
 import android.widget.ScrollView;
 import android.widget.TextView;
 
@@ -23,7 +20,6 @@ import java.util.Collections;
 
 import io.noties.markwon.AbstractMarkwonPlugin;
 import io.noties.markwon.BlockHandlerDef;
-import io.noties.markwon.LinkResolverDef;
 import io.noties.markwon.Markwon;
 import io.noties.markwon.MarkwonConfiguration;
 import io.noties.markwon.MarkwonSpansFactory;
@@ -31,7 +27,6 @@ import io.noties.markwon.MarkwonVisitor;
 import io.noties.markwon.SoftBreakAddsNewLinePlugin;
 import io.noties.markwon.core.CoreProps;
 import io.noties.markwon.core.MarkwonTheme;
-import io.noties.markwon.core.spans.HeadingSpan;
 import io.noties.markwon.core.spans.LastLineSpacingSpan;
 import io.noties.markwon.image.ImageItem;
 import io.noties.markwon.image.ImagesPlugin;
@@ -62,7 +57,9 @@ public class BasicPluginsActivity extends ActivityWithMenuOptions {
                 .add("headingNoSpace", this::headingNoSpace)
                 .add("headingNoSpaceBlockHandler", this::headingNoSpaceBlockHandler)
                 .add("allBlocksNoForcedLine", this::allBlocksNoForcedLine)
-                .add("anchor", this::anchor);
+                .add("anchor", this::anchor)
+                .add("letterOrderedList", this::letterOrderedList)
+                .add("tableOfContents", this::tableOfContents);
     }
 
     @Override
@@ -323,26 +320,26 @@ public class BasicPluginsActivity extends ActivityWithMenuOptions {
     }
 
     private void headingNoSpaceBlockHandler() {
-final Markwon markwon = Markwon.builder(this)
-        .usePlugin(new AbstractMarkwonPlugin() {
-            @Override
-            public void configureVisitor(@NonNull MarkwonVisitor.Builder builder) {
-                builder.blockHandler(new BlockHandlerDef() {
+        final Markwon markwon = Markwon.builder(this)
+                .usePlugin(new AbstractMarkwonPlugin() {
                     @Override
-                    public void blockEnd(@NonNull MarkwonVisitor visitor, @NonNull Node node) {
-                        if (node instanceof Heading) {
-                            if (visitor.hasNext(node)) {
-                                visitor.ensureNewLine();
-                                // ensure new line but do not force insert one
+                    public void configureVisitor(@NonNull MarkwonVisitor.Builder builder) {
+                        builder.blockHandler(new BlockHandlerDef() {
+                            @Override
+                            public void blockEnd(@NonNull MarkwonVisitor visitor, @NonNull Node node) {
+                                if (node instanceof Heading) {
+                                    if (visitor.hasNext(node)) {
+                                        visitor.ensureNewLine();
+                                        // ensure new line but do not force insert one
+                                    }
+                                } else {
+                                    super.blockEnd(visitor, node);
+                                }
                             }
-                        } else {
-                            super.blockEnd(visitor, node);
-                        }
+                        });
                     }
-                });
-            }
-        })
-        .build();
+                })
+                .build();
 
         final String md = "" +
                 "# Title title title title title title title title title title \n\ntext text text text";
@@ -384,85 +381,6 @@ final Markwon markwon = Markwon.builder(this)
         markwon.setMarkdown(textView, md);
     }
 
-//    public void step_6() {
-//
-//        final Markwon markwon = Markwon.builder(this)
-//                .usePlugin(HtmlPlugin.create())
-//                .usePlugin(new AbstractMarkwonPlugin() {
-//                    @Override
-//                    public void configure(@NonNull Registry registry) {
-//                        registry.require(HtmlPlugin.class, plugin -> plugin.addHandler(new SimpleTagHandler() {
-//                            @Override
-//                            public Object getSpans(@NonNull MarkwonConfiguration configuration, @NonNull RenderProps renderProps, @NonNull HtmlTag tag) {
-//                                return new AlignmentSpan.Standard(Layout.Alignment.ALIGN_CENTER);
-//                            }
-//
-//                            @NonNull
-//                            @Override
-//                            public Collection<String> supportedTags() {
-//                                return Collections.singleton("center");
-//                            }
-//                        }));
-//                    }
-//                })
-//                .build();
-//    }
-
-    // text lifecycle (after/before)
-    // rendering lifecycle (before/after)
-    // renderProps
-    // process
-
-    private static class AnchorSpan {
-        final String anchor;
-
-        AnchorSpan(@NonNull String anchor) {
-            this.anchor = anchor;
-        }
-    }
-
-    @NonNull
-    private String createAnchor(@NonNull CharSequence content) {
-        return String.valueOf(content)
-                .replaceAll("[^\\w]", "")
-                .toLowerCase();
-    }
-
-    private static class AnchorLinkResolver extends LinkResolverDef {
-
-        interface ScrollTo {
-            void scrollTo(@NonNull View view, int top);
-        }
-
-        private final ScrollTo scrollTo;
-
-        AnchorLinkResolver(@NonNull ScrollTo scrollTo) {
-            this.scrollTo = scrollTo;
-        }
-
-        @Override
-        public void resolve(@NonNull View view, @NonNull String link) {
-            if (link.startsWith("#")) {
-                final TextView textView = (TextView) view;
-                final Spanned spanned = (Spannable) textView.getText();
-                final AnchorSpan[] spans = spanned.getSpans(0, spanned.length(), AnchorSpan.class);
-                if (spans != null) {
-                    final String anchor = link.substring(1);
-                    for (AnchorSpan span: spans) {
-                        if (anchor.equals(span.anchor)) {
-                            final int start = spanned.getSpanStart(span);
-                            final int line = textView.getLayout().getLineForOffset(start);
-                            final int top = textView.getLayout().getLineTop(line);
-                            scrollTo.scrollTo(textView, top);
-                            return;
-                        }
-                    }
-                }
-            }
-            super.resolve(view, link);
-        }
-    }
-
     private void anchor() {
         final String lorem = getString(R.string.lorem);
         final String md = "" +
@@ -472,32 +390,46 @@ final Markwon markwon = Markwon.builder(this)
                 lorem;
 
         final Markwon markwon = Markwon.builder(this)
-                .usePlugin(new AbstractMarkwonPlugin() {
-                    @Override
-                    public void configureConfiguration(@NonNull MarkwonConfiguration.Builder builder) {
-                        builder.linkResolver(new AnchorLinkResolver((view, top) -> scrollView.smoothScrollTo(0, top)));
-                    }
+                .usePlugin(new AnchorHeadingPlugin((view, top) -> scrollView.smoothScrollTo(0, top)))
+                .build();
 
-                    @Override
-                    public void afterSetText(@NonNull TextView textView) {
-                        final Spannable spannable = (Spannable) textView.getText();
-                        // obtain heading spans
-                        final HeadingSpan[] spans = spannable.getSpans(0, spannable.length(), HeadingSpan.class);
-                        if (spans != null) {
-                            for (HeadingSpan span : spans) {
-                                final int start = spannable.getSpanStart(span);
-                                final int end = spannable.getSpanEnd(span);
-                                final int flags = spannable.getSpanFlags(span);
-                                spannable.setSpan(
-                                        new AnchorSpan(createAnchor(spannable.subSequence(start, end))),
-                                        start,
-                                        end,
-                                        flags
-                                );
-                            }
-                        }
-                    }
-                })
+        markwon.setMarkdown(textView, md);
+    }
+
+    private void letterOrderedList() {
+        // bullet list nested in ordered list renders letters instead of bullets
+        final String md = "" +
+                "1. Hello there!\n" +
+                "1. And here is how:\n" +
+                "   - First\n" +
+                "   - Second\n" +
+                "   - Third\n" +
+                "      1. And first here\n\n";
+
+        final Markwon markwon = Markwon.builder(this)
+                .usePlugin(new BulletListIsOrderedWithLettersWhenNestedPlugin())
+                .build();
+
+        markwon.setMarkdown(textView, md);
+    }
+
+    private void tableOfContents() {
+        final String lorem = getString(R.string.lorem);
+        final String md = "" +
+                "# First\n" +
+                "" + lorem + "\n\n" +
+                "# Second\n" +
+                "" + lorem + "\n\n" +
+                "## Second level\n\n" +
+                "" + lorem + "\n\n" +
+                "### Level 3\n\n" +
+                "" + lorem + "\n\n" +
+                "# First again\n" +
+                "" + lorem + "\n\n";
+
+        final Markwon markwon = Markwon.builder(this)
+                .usePlugin(new TableOfContentsPlugin())
+                .usePlugin(new AnchorHeadingPlugin((view, top) -> scrollView.smoothScrollTo(0, top)))
                 .build();
 
         markwon.setMarkdown(textView, md);